moved error string classification in archival
