+++ internal/engine/netx/errorx/errorx_test.go
@@ -1,212 +1,8 @@
 package errorx
 
 import (
-	"context"
-	"crypto/tls"
-	"crypto/x509"
-	"errors"
-	"io"
-	"net"
-	"syscall"
 	"testing"
-
-	"github.com/google/go-cmp/cmp"
-	"github.com/lucas-clemente/quic-go"
-	"github.com/pion/stun"
 )
-
-func TestMaybeBuildFactory(t *testing.T) {
-	err := SafeErrWrapperBuilder{
-		ConnID:        1,
-		DialID:        10,
-		Error:         errors.New("mocked error"),
-		TransactionID: 100,
-	}.MaybeBuild()
-	var target *ErrWrapper
-	if errors.As(err, &target) == false {
-		t.Fatal("not the expected error type")
-	}
-	if target.ConnID != 1 {
-		t.Fatal("wrong ConnID")
-	}
-	if target.DialID != 10 {
-		t.Fatal("wrong DialID")
-	}
-	if target.Failure != "unknown_failure: mocked error" {
-		t.Fatal("the failure string is wrong")
-	}
-	if target.TransactionID != 100 {
-		t.Fatal("the transactionID is wrong")
-	}
-	if target.WrappedErr.Error() != "mocked error" {
-		t.Fatal("the wrapped error is wrong")
-	}
-}
-
-func TestToFailureString(t *testing.T) {
-	t.Run("for already wrapped error", func(t *testing.T) {
-		err := SafeErrWrapperBuilder{Error: io.EOF}.MaybeBuild()
-		if toFailureString(err) != FailureEOFError {
-			t.Fatal("unexpected result")
-		}
-	})
-	t.Run("for ErrDNSBogon", func(t *testing.T) {
-		if toFailureString(ErrDNSBogon) != FailureDNSBogonError {
-			t.Fatal("unexpected result")
-		}
-	})
-	t.Run("for context.Canceled", func(t *testing.T) {
-		if toFailureString(context.Canceled) != FailureInterrupted {
-			t.Fatal("unexpected result")
-		}
-	})
-	t.Run("for x509.HostnameError", func(t *testing.T) {
-		var err x509.HostnameError
-		if toFailureString(err) != FailureSSLInvalidHostname {
-			t.Fatal("unexpected result")
-		}
-	})
-	t.Run("for x509.UnknownAuthorityError", func(t *testing.T) {
-		var err x509.UnknownAuthorityError
-		if toFailureString(err) != FailureSSLUnknownAuthority {
-			t.Fatal("unexpected result")
-		}
-	})
-	t.Run("for x509.CertificateInvalidError", func(t *testing.T) {
-		var err x509.CertificateInvalidError
-		if toFailureString(err) != FailureSSLInvalidCertificate {
-			t.Fatal("unexpected result")
-		}
-	})
-	t.Run("for operation was canceled error", func(t *testing.T) {
-		if toFailureString(errors.New("operation was canceled")) != FailureInterrupted {
-			t.Fatal("unexpected result")
-		}
-	})
-	t.Run("for EOF", func(t *testing.T) {
-		if toFailureString(io.EOF) != FailureEOFError {
-			t.Fatal("unexpected results")
-		}
-	})
-	t.Run("for connection_refused", func(t *testing.T) {
-		if toFailureString(syscall.ECONNREFUSED) != FailureConnectionRefused {
-			t.Fatal("unexpected results")
-		}
-	})
-	t.Run("for connection_reset", func(t *testing.T) {
-		if toFailureString(syscall.ECONNRESET) != FailureConnectionReset {
-			t.Fatal("unexpected results")
-		}
-	})
-	t.Run("for connection_reset wrapped", func(t *testing.T) {
-		if toFailureString(&net.OpError{Err: syscall.ECONNRESET}) != FailureConnectionReset {
-			t.Fatal("unexpected results", toFailureString(&net.DNSConfigError{Err: syscall.ECONNRESET}))
-		}
-	})
-	t.Run("for context deadline exceeded", func(t *testing.T) {
-		ctx, cancel := context.WithTimeout(context.Background(), 1)
-		defer cancel()
-		<-ctx.Done()
-		if toFailureString(ctx.Err()) != FailureGenericTimeoutError {
-			t.Fatal("unexpected results")
-		}
-	})
-	t.Run("for stun's transaction is timed out", func(t *testing.T) {
-		if toFailureString(stun.ErrTransactionTimeOut) != FailureGenericTimeoutError {
-			t.Fatal("unexpected results")
-		}
-	})
-	t.Run("for i/o error", func(t *testing.T) {
-		ctx, cancel := context.WithTimeout(context.Background(), 1)
-		defer cancel() // fail immediately
-		conn, err := (&net.Dialer{}).DialContext(ctx, "tcp", "www.google.com:80")
-		if err == nil {
-			t.Fatal("expected an error here")
-		}
-		if conn != nil {
-			t.Fatal("expected nil connection here")
-		}
-		if toFailureString(err) != FailureGenericTimeoutError {
-			t.Fatal("unexpected results")
-		}
-	})
-	t.Run("for TLS handshake timeout error", func(t *testing.T) {
-		err := errors.New("net/http: TLS handshake timeout")
-		if toFailureString(err) != FailureGenericTimeoutError {
-			t.Fatal("unexpected results")
-		}
-	})
-	t.Run("for no such host", func(t *testing.T) {
-		if toFailureString(&net.DNSError{
-			Err: "no such host",
-		}) != FailureDNSNXDOMAINError {
-			t.Fatal("unexpected results")
-		}
-	})
-	t.Run("for errors including IPv4 address", func(t *testing.T) {
-		input := errors.New("read tcp 10.0.2.15:56948->93.184.216.34:443: use of closed network connection")
-		expected := "unknown_failure: read tcp [scrubbed]->[scrubbed]: use of closed network connection"
-		out := toFailureString(input)
-		if out != expected {
-			t.Fatal(cmp.Diff(expected, out))
-		}
-	})
-	t.Run("for errors including IPv6 address", func(t *testing.T) {
-		input := errors.New("read tcp [::1]:56948->[::1]:443: use of closed network connection")
-		expected := "unknown_failure: read tcp [scrubbed]->[scrubbed]: use of closed network connection"
-		out := toFailureString(input)
-		if out != expected {
-			t.Fatal(cmp.Diff(expected, out))
-		}
-	})
-	// QUIC failures
-	t.Run("for connection_refused", func(t *testing.T) {
-		if toFailureString(&quic.TransportError{ErrorCode: quic.ConnectionRefused}) != FailureConnectionRefused {
-			t.Fatal("unexpected results")
-		}
-	})
-	t.Run("for connection_reset", func(t *testing.T) {
-		if toFailureString(&quic.StatelessResetError{}) != FailureConnectionReset {
-			t.Fatal("unexpected results")
-		}
-	})
-	t.Run("for incompatible quic version", func(t *testing.T) {
-		if toFailureString(&quic.VersionNegotiationError{}) != FailureNoCompatibleQUICVersion {
-			t.Fatal("unexpected results")
-		}
-	})
-	t.Run("for i/o error", func(t *testing.T) {
-		ctx, cancel := context.WithTimeout(context.Background(), 1)
-		defer cancel() // fail immediately
-		udpAddr := &net.UDPAddr{IP: net.ParseIP("216.58.212.164"), Port: 80, Zone: ""}
-		udpConn, err := net.ListenUDP("udp", &net.UDPAddr{IP: net.IPv4zero, Port: 0})
-		if err != nil {
-			t.Fatal(err)
-		}
-		sess, err := quic.DialEarlyContext(ctx, udpConn, udpAddr, "google.com:80", &tls.Config{}, &quic.Config{})
-		if err == nil {
-			t.Fatal("expected an error here")
-		}
-		if sess != nil {
-			t.Fatal("expected nil session here")
-		}
-		if toFailureString(err) != FailureGenericTimeoutError {
-			t.Fatal("unexpected results")
-		}
-	})
-	t.Run("for QUIC handshake timeout error", func(t *testing.T) {
-		err := &quic.HandshakeTimeoutError{}
-		if toFailureString(err) != FailureGenericTimeoutError {
-			t.Fatal("unexpected results")
-		}
-	})
-	t.Run("for QUIC connection timeout error", func(t *testing.T) {
-		err := &quic.IdleTimeoutError{}
-		if toFailureString(err) != FailureGenericTimeoutError {
-			t.Fatal("unexpected results")
-		}
-	})
-}
 
 func TestToOperationString(t *testing.T) {
 	t.Run("for connect", func(t *testing.T) {
